ecom/requirements.txt
Django==4.2.3
django-allauth==0.63.3
django-anymail==10.3
django-celery-results==2.5.1
email-validator==1.3.1
exceptiongroup==1.1.3
execnet==2.0.2
executing==1.2.0
Faker==25.9.2
fastjsonschema==2.19.0
GitPython==3.1.42
glob2==0.7
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.60.0
oauthlib==3.2.2
openai==1.35.7
path==16.7.1
path.py==12.5.0
Pillow==9.4.0
pipenv==2023.12.1
platformdirs==3.9.1
pluggy==1.3.0
psycopg2-binary==2.9.7
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
roboflow==1.1.25

Send2Trash==1.8.2
shapely==2.0.3
six==1.16.0
sklearn==0.0.post7
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
SQLAlchemy==1.4.46

stripe==9.12.0

vine==5.1.0
virtualenv==20.25.1

wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0

WTForms==3.0.1

